requirements.txt
rdkit-pypi
gdown
OpenNMT-py==2.2.0